Ajusta Campos 'start' e 'stop' módulo 'mail_activity_done'
